refactor: UpdateLocationUseCase에 ETA 업데이트 기능 통합
- UpdateLocationAndEtaUseCase 제거 (불필요한 중복)
- UpdateLocationUseCase를 확장하여 선택적으로 Route 파라미터 받도록 수정
- 중복 코드 제거 (shouldUpdate, calculateDistance)
- BroadcastLocationUseCase에서 통합된 UpdateLocationUseCase 사용
- 하위 호환성 유지: invoke(location)과 invoke(location, route) 모두 지원
